requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
streamlit>=1.22.0
matplotlib>=3.4.0
